Add ul4 embedded grammar
It basically wraps the default grammar and applies it inside string to improve visibility. At the moment the new grammer is using the `text.html.basic` from the old one, but this can be changed to keep the basic grammar clean and to allow the ability to add more wrapping languages.